`fn rav1d_itx_dsp_init_*`: Deduplicate w/ generics
